【#】用户和角色权重类型修改；

+++ yiboweb-security/src/main/java/com/yibo/modules/base/entity/Role.java
@@ -45,7 +45,7 @@ public class Role extends DataEntity<String>{
     private String roleCode;
     
     @ApiModelProperty(value = "角色权重（倒序）")
-    private Double roleWeight;
+    private Integer roleWeight;
     
     @ApiModelProperty(value = "角色分类（预留）")
     private String roleType;
